Update and rename auto.ps1 to AutoDelete.ps1

+++ AutoDelete.ps1
@@ -17,5 +17,11 @@ $prurl = $PullRequest.url
 $prinfo = curl -X GET -u ${Owner}:$TokenGithub $prurl | ConvertFrom-Json
 $branch = $prinfo.head.ref
 
-curl -X DELETE -u ${Owner}:$TokenGithub https://api.github.com/repos/$Owner/$Repo/git/refs/heads/$branch
-write-host "$branch be deleted"
+if($branch -ne 'development'){
+  write-Host 'Deleting branch...'
+  curl -X DELETE -u ${Owner}:$TokenGithub https://api.github.com/repos/$Owner/$Repo/git/refs/heads/$branch
+  write-Host '$branch branch deleted'
+}
+else{
+  write-host "This Branch can't be deleted"
+}
